+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/AdaPots.hs
@@ -33,7 +33,6 @@ import Cardano.Ledger.Compactible (fromCompact)
 import Cardano.Ledger.Core
 import Cardano.Ledger.Shelley.Governance (EraGov (..))
 import Cardano.Ledger.Shelley.LedgerState.Types (
-  AccountState (..),
   EpochState (..),
   LedgerState (..),
   UTxOState (..),
@@ -41,7 +40,7 @@ import Cardano.Ledger.Shelley.LedgerState.Types (
   utxosGovStateL,
  )
 import Cardano.Ledger.Shelley.TxBody (unWithdrawals)
-import Cardano.Ledger.State (UTxO (..), coinBalance, txInsFilter, txouts)
+import Cardano.Ledger.State
 import Cardano.Ledger.UMap (sumRewardsUView)
 import Control.DeepSeq (NFData)
 import Data.Foldable (fold)